feat(select): Makes SelectUpdateStatus more broad and accepts different mutations
